[Refactor] Move mean pooling to pool family
