+++ tileops/kernels/pool/__init__.py
@@ -4,6 +4,7 @@ from .avg_pool3d import AvgPool3dKernel, AvgPool3dSpatialKernel
 from .max_pool1d import MaxPool1dKernel, MaxPool1dWithIndicesKernel
 from .max_pool2d import MaxPool2dKernel, MaxPool2dWithIndicesKernel
 from .max_pool3d import MaxPool3dKernel, MaxPool3dWithIndicesKernel
+from .mean_pooling import MeanPoolingFwdKernel
 
 __all__ = [
     "AvgPool1dKernel",
@@ -18,4 +19,5 @@ __all__ = [
     "MaxPool2dWithIndicesKernel",
     "MaxPool3dKernel",
     "MaxPool3dWithIndicesKernel",
+    "MeanPoolingFwdKernel",
 ]
