+++ workloads/mean_pooling.py
